Split up named_pipe.c in win and unix
