+++ src/Makevars
@@ -1,5 +1,5 @@
 OBJECTS = init.o unix/childlist.o unix/connection.o unix/poll.o \
-          unix/processx.o unix/sigchld.o unix/utils.o
+          unix/processx.o unix/sigchld.o unix/utils.o unix/named_pipe.o
 
 .PHONY: all clean
